\name{auto.optimize}
\alias{auto.optimize}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Optimize rflasso models by pruning the trees
}
\description{
Optimize the Random Forest tree by pruning it. This function utilizes out of bag error to find the optimal model.
}
\usage{
auto.optimize(bagged.model,step.size=.025,guess=1.1,number.of.cores)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{bagged.model}{Compatible with models generated by the "build.tree" function.}
  \item{step.size}{Step size that the "significance" parameter changes.}
  \item{guess}{Initial guess of the "significance" parameter.}
  \item{number.of.cores}{Number of cores to be utilized.}
}
\details{
This function prunes the trees based on the "significance" parameter. The "significance" parameter is multiplied by the MSE of the terminal nodes and compared to the MSE of the parent node. If the penalized MSE of the terminal nodes is greater than the parent node, the leaves are pruned. After the trees have been pruned, the out of bag error is calculated. The "significance" parameter is initialized by the "guess" argument. The significance parameter changes at each iteration based on step.size. The Random Forest with the minimum out.of.bag error is returned. This function always uses model="lasso".
}
\value{
Returns the optimal model object.
}

\author{
Alexander Wendelborn
apwendelborn@gmail.com
}


\seealso{
build.tree, predict.lm.from.rf, out.of.bag.error
}

