\name{build.tree}
\alias{build.tree}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Generate a random forest ensemble which supports lasso regression after pooling the terminal leaves
}
\description{
Builds a standard random forest with the help from parallel processing on multicore machines. The model object generated by this function is compatible with the functions which generate predictions by pooling observations from all trees and fitting a linear lasso model. Please note: this package only support regression currently.
}
\usage{
build.tree(training.data,dependent.variable,independent.variables,samples.per.tree,
minimum.leaf.size=2,number.of.independent.variables.to.sample,number.of.cores,
number.of.trees,index.columns=NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{training.data}{Accepts R data.frames as an input.}
  \item{dependent.variable}{Name of column from training.data with the dependent variable. Only supports regression currently.}
  \item{independent.variables}{Vector of names of columns from training.data with the independent variables.}
  \item{samples.per.tree}{Number of samples per tree in the random forest ensemble.}
  \item{minimum.leaf.size}{Minimum size of terminal leaves.}
  \item{number.of.independent.variables.to.sample}{Number of variables to sample at each split.}
  \item{number.of.cores}{Number of cores to use when fitting the random forest.}
  \item{number.of.trees}{Number of trees to build for the random forest ensemble.}
  \item{index.columns}{Useful for labelling predictions. These columns are not used in fitting the random forest and should uniquely identify samples.}
}
\value{
Returns a model object which is can be used to make predictions.
}
\references{
Breiman, L. (2001), Random Forests, Machine Learning 45(1), 5-32.
}
\author{
Alexander Wendelborn
<apwendelborn@gmail.com>
}


\seealso{
predict.lm.from.rf, out.of.bag.error, auto.optimize
}
