\name{out.of.bag.error}
\alias{out.of.bag.error}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Out of Bag Error for rflasso
}
\description{
Calculate the aggregate RMSE for each sample in the training dataset by using the out of bag random forest. This function supports either using lasso on the pooled leaves or simple averaging.
}
\usage{
out.of.bag.error(bagged.model,model="lasso",number.of.cores)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{bagged.model}{Compatible with models generated by the "build.tree" function.}
  \item{model}{Either "lasso" or "average".}
  \item{number.of.cores}{Number of cores to be utilized.}
}

\value{
This function will return the RMSE of the predictions versus the actual values.
}

\author{
Alexander Wendelborn
<apwendelborn@gmail.com>
}

\note{
Lasso is tuned by GCV using the "msgps" package. This function is useful for determining whether the standard random forest predictions or lasso based on pooled leaves will perform better on out of sample data.
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
build.tree, predict.lm.from.rf, auto.optimize
}

